Modified to do explicit type casting

+++ Runtime/Core/Window.h
@@ -17,7 +17,7 @@ namespace Mile
 
       unsigned int GetResWidth( ) const { return m_resWidth; }
       unsigned int GetResHeight( ) const { return m_resHeight; }
-      Vector2 GetResolution( ) const { return Vector2( m_resWidth, m_resHeight ); }
+      Vector2 GetResolution( ) const { return Vector2( static_cast<float>( m_resWidth ), static_cast<float>( m_resHeight ) ); }
 
       Float GetRefreshRate( ) const { return m_refreshRate; }
